requirements.txt
Django-UserService==1.0.2